Increase strictness of Mockito tests
